fix: ロール ID を列の実型 integer に揃え 16bit への縮小をやめる
roles.id / users.role_id を本番の実型 integer に直した結果、sqlc が int32 を
生成するようになり、domain 側の uint16 へ代入する箇所が 32bit → 16bit の
縮小変換になっていた。ロールは 1〜3 の固定採番なので実害は出ていないが、
エラーも出さずに値が化ける型の穴なので塞ぐ。

domain の Role.ID / RoleID 定数 / User.RoleID を int32 にし、変換自体を無くす。
不要になった int32(...) 6 箇所も除去（unconvert が検出）。

+++ backend/internal/adapter/persistence/company_learning_activity_repository.go
@@ -39,7 +39,7 @@ func (r *companyLearningActivityRepository) ListMemberActivities(
 		FromDate:  from,
 		CompanyID: cid,
 		// trainee 判定は正規化後の正である role_id で行う（FRESTYLE-311）。
-		TraineeRoleID: int32(domain.RoleIDTrainee),
+		TraineeRoleID: domain.RoleIDTrainee,
 	})
 	if err != nil {
 		return nil, err
